[Tizen] Call OnBake(), OnSet only if the value is really changed
Let we check the current value and inputed value, so if they are same, ignore.

It will be useful when we don't want to mark dirty for animatable property doesn't changed.

For performance issue, let we just check only for bool / int / float.

TODO : Chould we check them also at Vector2,3,4?

Change-Id: I70b36eab97585a09c2c021da4f37a1d983a771af
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>