Updating framebuffer to use renderpass
Vulkan framebuffer requires a renderpass to be defined before
the framebuffer can be created. Currently in gles backend, this
order is irrelevant, and is not defined in the graphics api.

Added Graphics::RenderPass vector to the
Graphics::FramebufferCreateInfo - this vector may have more than one
compatible render pass in it. The first render pass will be used to
create the framebuffer, but BeginRenderPass may use any of the defined
render passes with that framebuffer (as long as they are compatible in
vulkan, i.e. only differ by layout or load/store ops).

Change-Id: I66ce8530a34bcb1da57461db5644771efe0af48c