Added clip transform matrix to the Graphics Controller
For GL, this matrix should not be used, so doesn't affect
render time. For Vulkan, this will change to Vulkan
clip space, so adds another matrix multiply.

(Consider moving the projection matrix setup to graphics
controller...)

Moved the viewport/scissor setting to the secondary command buffer.
(It's not an issue for GL, but is for Vulkan)

Also ensured that the viewport near/far clip values are set.

Change-Id: I4c1842fad5766d9be769fe9dce79386f84b5459a