Do not update componentwise animatable property's Set/Bake
We don't update PropertyOwner's update state when the value was
not actuall changed.

But it doesn't reset the dirty flag of AnimatablePropertyBase.

Due to the AnimatablePropertyBase is not a clean flag,
the renderer who use constrait become always check as dirty
when we check dirty rect.

To fix this kind of issue, let we don't call OnSet and OnBake
if the componentwise value is equal.
(Since float comparision is cheap enough)

Change-Id: I3b3744aa420049523ece5bb72b915c8dd192d4f3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>