Changed Matrix3 uniform write to use stride
If the uniform is in a block, then each row of 3 is padded
to 4 floats(16 bytes). Can get the stride from GLES/Vulkan reflection.

So, use matrixStride if not in the standalone uniform block
(which means it's gles2 and packed tightly).

Added test cases to check Mat3 uniform writing in both standalone
and uniform buffer.

Change-Id: If6e7b49318dae795972d5ae76092df48f4713cab
Signed-off-by: David Steele <david.steele@samsung.com>