[Tizen] Make WeakHandleBase::mImpl as unique_ptr + nullcheck for Reset and comparision
Until now, WeakHandleBase::Reset() function and == operator
don't check mImpl was nullptr.

But it could be nullptr when we use move operations.

So we should check nullptr before call mImpl->Reset() and operator==.

===

Also, to avoid this kind of leak, let we use std::unique_ptr instead of raw pointer.

Change-Id: Icd0b603b825d6a1f5eb7697466da98af05540b2b
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>