Support Property::Value::GetHash() + Implement lite equality operator for Map and Array
Let we support to get hash value of Property
+
Seperate the hash calculation function as internal,
so we can remove duplicated hash calculation logics in dali-core.
+
Support equality operasion for Property::Map and Property::Array.

TODO : For now, we don't consider Eplison when we check equality.
TODO : For now, we only compare hash value, not an exact value of array and map.
It should be supported in future.

Change-Id: Ie0e19f46f35cf98d84ed9dda0c3dd0fca4a8dd61
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>