[Tizen] Support Property::Value::GetHash() + Implement lite equality operator for Map and Array
Let we support to get hash value of Property
+
Seperate the hash calculation function as internal,
so we can remove duplicated hash calculation logics in dali-core.
+
Support equality operasion for Property::Map and Property::Array.

TODO : For now, we don't consider Eplison when we check equality.
TODO : For now, we only compare hash value, not an exact value of array and map.
It should be supported in future.

Allow to use moved Property::Map & Property::Array

Until now, we make those value's mImpl as nullptr.
So their was a lots of DALI_ASSERT_DEBUG() to check
whether they are moved or not.

Those codes might not be fair than other moved objects implementation.

we allow to use moved object like WeakHandle, or Dali::Vector.
Let we also allow to use them.

And let we make surely assert by DALI_ASSERT_ALWAYS if we need, instead DALI_ASSERT_DEBUG

Change-Id: If08853a1635a8c33317edc1d25531223f7c427ed
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>