Make PixelData Stride return byte size, instead of pixel
Some cases doesn't have stride byte as multiply of BytesPerPixel().
But until now, PixelBuffer and PixelData only support Stride as pixels count
(It looks since GL_UNPACK_ROW_LENGTH is pixel scale, not a bytes.)

For example, If we have 2 width image with RGB888 format, stride byte
should be 8 byte, which is not a multiply of 3 byte.

To support those cases, let we set stride value as bytes scale.

Change-Id: I27aa9757e7320475b7f0519a62262d1ae2f7e07b
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>