Add Rendered flag for PreRender the Scene
Let we make some interface whether given scene rendered previous frame or not.

Until now, we only had "current" scene has render instruction to scene or not.
But also, we should clear the scene if "previous" scene had render instruction
but not now.

Since it could not be known by single boolean value, we make another class
to communicate the PreRender result.

By this infomation, we should determine whether we should call
ClearScene() or not even if fullSwap keyword not applied.

Furthermore, Let we make PreRender(Scene) API name as PreRenderScene,
pairwise with RenderScene().

Change-Id: Iab820942aff1d136cef1ab111901dc99c7ccdcf7
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>