Ensure to call program.GetUniform() only 1 times even if we found failed.
Until now, UniformIndexMap state mark as initialized = false if given unform
not found at shader.

In this case, we always try to re-search about given uniform name, which is useless.

To avoid this case, let we make the uniform found result as enum,
and so we can ensure that program.GetUniform only 1 time.

Change-Id: I23a180980f556e3f05302622453258f87742344d
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>