(Constraint) Implement SetApplyRate(APPLY_ONCE)
Let we implement the way to apply constraint only 1 time.

Follow with RenderTask::SetRefreshRate(), we also allow to seperate
the ratio of constraint applying.

TODO : We should seperate the container of constraints at PropertyOwner
which is for ApplyOnce or not.

For now, let we just implement it as pure-way for PoC

Change-Id: I0a1073ec3aaf899f7d010c7ae471d2d5006c1a65
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>