Fix memory leaks in SignalDelegate.
The issue is found in UtcDaliSignalDelegateIsConnectedP, etc by ASan.
When SignalDelegate's Connect function is called, CallbackBaseFunctor
is always allocated in heap, but it is not deallocated any longer.
FunctorDelegate allocates a new memory for CallbackBaseFunctor
with CallbackBaseFunctor copy constructor, then FunctorDelegate only
gains ownership of CallbackBaseFunctor's mCallback.
So the memory of CallbackBaseFunctor instance in SignalDelegate's Connect
is leaked.

This patch is to make CallbackBaseFunctor movable to avoid memory leak.

Change-Id: I0f21da5c7b4139be305f00e93e6651585fd25ef9