Fix render target's extents on resize
On surface resize, the render target is not always
recreated, if the surface hasn't changed, for instance.
This means that the extents of the render target in it's
create info become out of date.

If these extents are then used later, the result is
inaccurate. (We are using them to calculate scissor
and viewport rects)

Add a mechanism to graphics controller to enable
the extents to be updated.

Change-Id: Ib2371e3657176cb017c52cab1c62ac4190ed6a66
Signed-off-by: David Steele <david.steele@samsung.com>